+++ autogen/variables.tf
@@ -86,6 +86,13 @@ variable "master_authorized_networks_config" {
   default = []
 }
 
+{% if private_cluster %}
+variable "enable_binary_authorization" {
+  description = "Enable BinAuthZ Admission controller"
+  default     = false
+}
+
+{% endif %}
 variable "horizontal_pod_autoscaling" {
   description = "Enable horizontal pod autoscaling addon"
   default     = true