+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Extending the adopted spec, each change should have a link to its corresponding 
 
 ### Added
 
+* Add configuration flag for enable BinAuthZ Admission controller [#160]
 * Support for a guest accelerator in node pool configuration. [#157]
 * Support to scale the default node cluster. [#149]
 * Support for configuring the network policy provider. [#159]